refactor（v0.2.1-optimal）：目录化 codec responses_codec
